requirements.txt
blinker==1.9.0
Brotli @ file:///C:/b/abs_3d36mno480/croot/brotli-split_1714483178642/work
certifi @ file:///C:/b/abs_1fw_exq1si/croot/certifi_1725551736618/work/certifi
charset-normalizer @ file:///croot/charset-normalizer_1721748349566/work
click==8.1.7
colorama==0.4.6
Deprecated==1.2.14
einops==0.6.0
filelock==3.16.1
Flask==3.0.3
Flask-Limiter==3.8.0
fsspec==2024.10.0
gunicorn==23.0.0
huggingface-hub==0.26.2
idna @ file:///C:/b/abs_aad84bnnw5/croot/idna_1714398896795/work
importlib_resources==6.4.5
imutils==0.5.4
itsdangerous==2.2.0
Jinja2==3.1.4
limits==3.13.0
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
mkl-fft==1.3.1
mkl-random @ file:///C:/ci_310/mkl_random_1643050563308/work
mkl-service==2.4.0
numpy @ file:///C:/b/abs_5ct9ex77k9/croot/numpy_and_numpy_base_1668593740598/work
opencv-python==4.6.0.66
ordered-set==4.1.0
packaging==24.2
pillow @ file:///C:/b/abs_32o8er3uqp/croot/pillow_1721059447598/work
Pygments==2.18.0
PySocks @ file:///C:/ci_310/pysocks_1642089375450/work
PyYAML==6.0.2
requests @ file:///C:/b/abs_c3508vg8ez/croot/requests_1731000584867/work
rich==13.9.4
six @ file:///tmp/build/80754af9/six_1644875935023/work
timm==0.6.12
torch==1.13.0
torchvision==0.14.0
tqdm==4.67.0
typing_extensions @ file:///C:/b/abs_0as9mdbkfl/croot/typing_extensions_1715268906610/work
urllib3 @ file:///C:/b/abs_9a_f8h_bn2/croot/urllib3_1727769836930/work
Werkzeug==3.1.3
win-inet-pton @ file:///C:/ci_310/win_inet_pton_1642658466512/work
wrapt==1.16.0
